read netmap updates from IPNBus rather than polling
This allows us to respond to node additions, removals, and renames in
near-realtime instead of waiting a minute for the next poll event.
Protect DNS entries map in a RWMutex, and replace the entire map on each
update to ensure old nodes are removed.

This also uses the Tailscale computed hostname for DNS entries, rather
than the reported hostname from the operating system. These can differ
if another host already exists on the tailnet with that name, or if an
admin has manually changed the name in the Tailscale admin panel. This
now ensures that we are using the same hostname that's actually used on
the tailnet.

This change also removes entries for nodes that have been shared into
the tailnet as well as Mullvad exit nodes. These hosts don't have unique
names within the tailnet and must be addressed using the FQDN for their
own tailnet. Because of that, I'm not sure that it ever makes sense to
add DNS entries for these nodes, but this leaves a TODO to reconsider
that in the future.