add support for using tsnet to connect to tailscale (#54)
This adds support for specifying a Tailscale auth key in the user's
Corefile. When provided, that key is used to have coredns join the
tailnet directly, rather than relying on having tailscale installed on
the local system. This simplifies deployment since everything needed is
compiled into the coredns binary, and allows Tailscale ACLs to be
applied to coredns specifically, independent of the hosting system.