enable the device web UI for tsnet (#84)
Since v1.56, Tailscale has a built-in web interface that can be used to
manage the device remotely over your tailnet. It is still subject to the
ACLs on the tailnet, so just enabling the interface (this change)
doesn't immediately expose anything. It just make it possible to
configure ACL access to remotely manage the device.

We recently made the same change to our golink service, as we believe it
is safe to enable this by default for these types of services:
https://github.com/tailscale/golink/pull/114

More info: https://tailscale.com/kb/1325/device-web-interface